Adaptação do código para usar as Classes Dto's e impletação da verificação se uma disciplina buscada existe

+++ src/main/java/com/dsc/student_social_network/services/CourseService.java
@@ -1,15 +1,18 @@
 package com.dsc.student_social_network.services;
 
+import com.dsc.student_social_network.dto.CommentDto;
 import com.dsc.student_social_network.dto.CourseDto;
+import com.dsc.student_social_network.dto.GradeDto;
 import com.dsc.student_social_network.entity.Comment;
 import com.dsc.student_social_network.entity.Course;
 import com.dsc.student_social_network.entity.Grade;
+import com.dsc.student_social_network.exception.CourseNotFoundException;
 import com.dsc.student_social_network.repository.CourseRepository;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
+import java.util.Comparator;
 import java.util.List;
 
 @Service
@@ -25,6 +28,9 @@ public class CourseService {
     }
 
     public CourseDto removeCourseById(int id) {
+        if (!courseRepository.existsById(id)) {
+            throw new CourseNotFoundException("Disciplina não encontrada", "CourseService.removeCourseById");
+        }
         Course course = courseRepository.findById(id).get();
         courseRepository.delete(course);
 
@@ -32,6 +38,9 @@ public class CourseService {
     }
 
     public CourseDto getCourseById(int id) {
+        if (!courseRepository.existsById(id)) {
+            throw new CourseNotFoundException("Disciplina não encontrada", "CourseService.getCourseById");
+        }
         return new CourseDto(courseRepository.findById(id).get());
     }
 
@@ -43,34 +52,50 @@ public class CourseService {
 
     public List<CourseDto> getAllCoursesByGradeDesc() {
         List<CourseDto> courseDtos = new ArrayList<>();
-        courseRepository.findAll(Sort.by(Sort.Direction.DESC, "grade"))
-                .forEach(course -> courseDtos.add(new CourseDto(course)));
+        courseRepository.findAll().forEach(course -> courseDtos.add(new CourseDto(course)));
+        courseDtos.sort(Comparator.comparing(a -> a.getGrade() * -1));
         return courseDtos;
     }
 
-    public CourseDto setNameCourseById(int id, String newName) {
+    public CourseDto setNameCourseById(int id, CourseDto courseDto) {
+        if (!courseRepository.existsById(id)) {
+            throw new CourseNotFoundException("Disciplina não encontrada", "CourseService.setNameCourseById");
+        }
         Course course = courseRepository.findById(id).get();
-        course.setName(newName);
+        course.setName(courseDto.getName());
         courseRepository.save(course);
         return new CourseDto(course);
     }
 
-    public CourseDto setGradeCourseById(int id, Double grade) {
+    public CourseDto setGradeCourseById(int id, GradeDto gradeDto) {
+        if (!courseRepository.existsById(id)) {
+            throw new CourseNotFoundException("Disciplina não encontrada", "CourseService.setGradeCourseById");
+        }
         Course course = courseRepository.findById(id).get();
-        Grade newGrade = new Grade(course, grade);
+        Grade newGrade = new Grade(course, gradeDto.getGrade());
         course.getGradeList().add(newGrade);
         return new CourseDto(courseRepository.save(course));
     }
 
-    public CourseDto addCommentToCourseById(int courseId, Comment comment) {
-        Course course = courseRepository.findById(courseId).get();
+    public CourseDto addCommentToCourseById(int id, CommentDto commentDto) {
+        if (!courseRepository.existsById(id)) {
+            throw new CourseNotFoundException("Disciplina não encontrada", "CourseService.addCommentToCourseById");
+        }
+        Course course = courseRepository.findById(id).get();
+        Comment comment = new Comment();
+        comment.setContent(commentDto.getContent());
+        comment.setCourse(course);
+        comment.setUserEmail(commentDto.getUserEmail());
         course.getCommentaries().add(comment);
         courseRepository.save(course);
         return new CourseDto(course);
     }
 
-    public CourseDto removeCommentToCourseById(int courseId, String commentId) {
-        Course course = courseRepository.findById(courseId).get();
+    public CourseDto removeCommentToCourseById(int id, String commentId) {
+        if (!courseRepository.existsById(id)) {
+            throw new CourseNotFoundException("Disciplina não encontrada", "CourseService.removeCommentToCourseById");
+        }
+        Course course = courseRepository.findById(id).get();
         for (Comment comment : course.getCommentaries()) {
             if (comment.getId() == Integer.parseInt(commentId))
                 comment.setIsRemoved(1);
